[#14] Test Case를 구현중 피드백 필요

+++ src/main/java/com/hello/kurly/products/v1/repository/ProductRepository.java
@@ -2,9 +2,10 @@ package com.hello.kurly.products.v1.repository;
 
 import com.hello.kurly.products.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
-import java.math.BigInteger;
-
-public interface ProductRepository extends JpaRepository<Product, BigInteger> {
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
 }
